Refactor to express-based Azure Function
