+++ package.json
@@ -8,7 +8,9 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
-    "@azure/functions": "^4.0.0"
+    "@azure/functions": "^4.0.0",
+    "azure-function-express": "^2.0.0",
+    "express": "^4.19.2"
   },
   "devDependencies": {},
   "keywords": [
