+++ src/handler.js
@@ -1,0 +1,4 @@
+const { createHandler } = require('azure-function-express');
+const app = require('./expressApp');
+
+module.exports = createHandler(app);
